// src/app/models/status.enum.ts
export enum StatusEnum {
    Pendente = 0,
    AnaliseCoordenador = 1,
    DocumentacaoPendenteEntrega = 2,
    DocumentacaoInvalida = 3,
    IndeferidoSemOferta = 4,
    IndeferidoOutros = 5,
    AprovacaoCandidato = 6,
    AguardandoAceiteAnalise = 7,
    ReanaliseCoordenador = 8,
    AnaliseAceita = 9,
    PreMatriculado = 10,
    Matriculado = 11,
    Aprovado = 12,
    Reprovado = 13,
    PendenteCorrecao = 14
}